# Conflicts:
#	decentralized-api/poc/orchestrator.go
#	docker-compose-local.yml
#	inference-chain/DockerfileGenesis
#	test2.sh

+++ inference-chain/x/inference/keeper/cost_calculator.go
@@ -4,11 +4,11 @@ import "github.com/productscience/inference/x/inference/types"
 
 const PerTokenCost = 1000
 
-func CalculateCost(inference types.Inference) uint64 {
+func CalculateCost(inference types.Inference) int64 {
 	// Very simple for now. Ultimately we will need to calculate this based on the model, and different
 	// values for completion and prompt
 	if inference.Status == types.InferenceStatus_STARTED {
-		return inference.MaxTokens * PerTokenCost
+		return int64(inference.MaxTokens * PerTokenCost)
 	}
-	return inference.CompletionTokenCount*PerTokenCost + inference.PromptTokenCount*PerTokenCost
+	return int64(inference.CompletionTokenCount*PerTokenCost + inference.PromptTokenCount*PerTokenCost)
 }
